+++ cms-20240330/alibabacloud_cms20240330/models/_get_digital_employee_response_body.py
@@ -14,6 +14,7 @@ class GetDigitalEmployeeResponseBody(DaraModel):
         default_rule: str = None,
         description: str = None,
         display_name: str = None,
+        employee_type: str = None,
         knowledges: main_models.GetDigitalEmployeeResponseBodyKnowledges = None,
         name: str = None,
         region_id: str = None,
@@ -26,6 +27,7 @@ class GetDigitalEmployeeResponseBody(DaraModel):
         self.default_rule = default_rule
         self.description = description
         self.display_name = display_name
+        self.employee_type = employee_type
         self.knowledges = knowledges
         self.name = name
         self.region_id = region_id
@@ -54,6 +56,9 @@ class GetDigitalEmployeeResponseBody(DaraModel):
 
         if self.display_name is not None:
             result['displayName'] = self.display_name
+
+        if self.employee_type is not None:
+            result['employeeType'] = self.employee_type
 
         if self.knowledges is not None:
             result['knowledges'] = self.knowledges.to_map()
@@ -88,6 +93,9 @@ class GetDigitalEmployeeResponseBody(DaraModel):
 
         if m.get('displayName') is not None:
             self.display_name = m.get('displayName')
+
+        if m.get('employeeType') is not None:
+            self.employee_type = m.get('employeeType')
 
         if m.get('knowledges') is not None:
             temp_model = main_models.GetDigitalEmployeeResponseBodyKnowledges()